Bring 4coder custom layer into this repo
Because programming the custom layer is integral to this project

+++ code/platform.h
@@ -24,7 +24,7 @@ struct ReadFileResult
   }
 };
 
-#define PLATFORM_WRITE_ENTIRE_FILE(NAME) b32 NAME(u8 *content, u64 content_size, char *filename)
+#define PLATFORM_WRITE_ENTIRE_FILE(NAME) b32 NAME(u8 *content, i32 content_size, char *filename)
 typedef PLATFORM_WRITE_ENTIRE_FILE(PlatformWriteEntireFile);
 
 #define PLATFORM_READ_ENTIRE_FILE(NAME) ReadFileResult NAME(Arena &arena, char *filename)
